bot/requirements.txt
# Discord
discord.py>=2.4.0

# Database
asyncpg>=0.30.0
